unique ir_generator_test.go
